+++ src/auth.rs
@@ -414,3 +414,12 @@ pub fn access_token<S: Into<String>>(con_token: &Token,
         try!(id.ok_or(error::Error::MissingValue("user_id"))),
         try!(username.ok_or(error::Error::MissingValue("screen_name")))))
 }
+
+///If the given tokens are valid, return the user information for the authenticated user.
+pub fn verify_tokens(con_token: &Token, access_token: &Token)
+    -> Result<Response<super::user::TwitterUser>, error::Error>
+{
+    let mut resp = try!(get(links::auth::VERIFY_CREDENTIALS, con_token, access_token, None));
+
+    parse_response(&mut resp)
+}