+++ src/lib.rs
@@ -78,5 +78,5 @@ pub mod entities;
 mod links;
 mod common;
 
-pub use auth::{Token, request_token, authorize_url, authenticate_url, access_token};
+pub use auth::{Token, request_token, authorize_url, authenticate_url, access_token, verify_tokens};
 pub use common::{Response, UserID, Cursor, CursorIter};